Add files via upload
Advanced conveyor has a two sensors one gate and one motor which runs in forward and reverse direction, there are 2 modes automatic and manual, automatic mode contains again three modes double sensor mode in which two sensors are active if any damage occurs to any of the sensors there is a single sensor mode in which either of the sensors can be used for normal operation of conveyor providing redundancy,
The motor start button is pressed then only the mode selection should take place
When ever stop button is pressed then all the process should stop releasing all the scada switches.
Two sensor mode 
In this mode two sensors are functional.
When the first sensor is actuated then the motor should run forward opening the gate at the same time
The motor is supposed to stop and gate to close when the second sensor is actuated showing the bag has left the conveyor
1st sensor mode.
In this mode 1st sensor is only active second sensor can be used for maintenance. 
If the first sensor is activated then the motor and gate operates for a pre-set time which can be set through scada.
The motor stops and gate closes after the pre-set time.
2nd sensor mode
In this mode 2nd sensor is only active 1st sensor can be used for maintenance.
The motor is activated by the status of the previous conveyor if the second sensor of the previous conveyor is activated then the motor runs with an extra second time duration than the pre-set time which is included for compensating the delay for the bag to reach the conveyor from the previous conveyor, this is done by adding an extra second in the math calculation of the conveyor.
If the previous conveyor’s second sensor is sensed the motor runs and gate opens for a pre-set time+1 second stops after the time duration or if the second sensor is sensed which ever is earlier.
Manual mode:
The motor can be operated manually in this mode
Motor forward and motor reverse can be run from scada.
If the motor is run either in forward or reverse direction then gate is supposed to open.
General
Stop button kills all the functions and releases all the buttons pressed.
If auto button is released then all the three modes under the auto is supposed to be released
If manual mode is released all the buttons under the manual is released.
If the active sensor in the respective mode actuates, then the count is increased by one hence the display in the scada represents the number of bags passing through the conveyor.

Program
All the time delays are generated by TT bit of a ton timer.
Extra time delay is generated using the same timer using a math function to increase the time delay.
Time is input in form of an integer is multiplied by 1000 to convert from seconds to milliseconds and is converted to time by a convert operation and fed to timer.
Scada interface is provided.
Mode selection is done by case of loop.
If else loop is used for other conditions.